// src/payload-types.ts
export interface Task {
  id: number;
  title: string;
  description?: string | null;
  status?: ('todo' | 'in-progress' | 'done') | null;
  updatedAt: string;
  createdAt: string;
}